requirement.txt
appdirs==1.4.3
AWSIoTPythonSDK==1.1.1
packaging==16.8
pyparsing==2.2.0
six==1.10.0
RPi.GPIO==0.6.3
pytz==2016.10
click==6.7